// AIRMAM5/Scripts/Models/Interface/Subject/KeyFrameCUModel.ts
/**新增,編輯關鍵影格 */
export interface KeyFrameCUModel {
    /**主題編號 */
    fsSUBJECT_ID: string;
    /**檔案編號 */
    fsFILE_NO: string;
    /** 關鍵影格標題 */
    Title: string;
    /**設定時間 */
    SetTime: string;
    /**關鍵影格描述 */
    Description: string;
}